+++ website/src/components/sandbox/sql/create-topic-tables/create-table.ts
@@ -11,7 +11,7 @@ export const createTableSqlTemplate = `
 			-- claim fence assumes ids are issued in INSERT order, and a cached
 			-- sequence hands out out-of-order id blocks
 
-			schema_version BIGINT NOT NULL,               -- the payload's version, from the producing Message type
+			schema_version INTEGER NOT NULL,              -- the payload's version, from the producing Message type
 			routing_key TEXT,
 			message_key TEXT,
 			compaction_rank BIGINT,                       -- NULL = this message never opted into compaction
